Reassure: refactor tests for more complex components
