+++ tests/perf-test/ChatFinderPage.perf-test.tsx
@@ -1,7 +1,6 @@
 import type * as NativeNavigation from '@react-navigation/native';
 import type {StackScreenProps} from '@react-navigation/stack';
-import {fireEvent, screen, waitFor} from '@testing-library/react-native';
-import type {TextMatch} from '@testing-library/react-native/build/matches';
+import {fireEvent, screen} from '@testing-library/react-native';
 import React, {useMemo} from 'react';
 import type {ComponentType} from 'react';
 import Onyx from 'react-native-onyx';
@@ -23,7 +22,6 @@ import type {Beta, PersonalDetails, Report} from '@src/types/onyx';
 import createCollection from '../utils/collections/createCollection';
 import createPersonalDetails from '../utils/collections/personalDetails';
 import createRandomReport from '../utils/collections/reports';
-import PusherHelper from '../utils/PusherHelper';
 import * as TestHelper from '../utils/TestHelper';
 import waitForBatchedUpdates from '../utils/waitForBatchedUpdates';
 import wrapOnyxWithWaitForBatchedUpdates from '../utils/wrapOnyxWithWaitForBatchedUpdates';
@@ -123,7 +121,6 @@ beforeEach(() => {
 // Clear out Onyx after each test so that each test starts with a clean state
 afterEach(() => {
     Onyx.clear();
-    PusherHelper.teardown();
 });
 
 type ChatFinderPageProps = StackScreenProps<RootStackParamList, typeof SCREENS.CHAT_FINDER_ROOT> & {
@@ -134,12 +131,14 @@ type ChatFinderPageProps = StackScreenProps<RootStackParamList, typeof SCREENS.C
 
 function ChatFinderPageWrapper(args: ChatFinderPageProps) {
     return (
-        <ComposeProviders components={[OnyxProvider, LocaleContextProvider, OptionListContextProvider, KeyboardStateProvider]}>
-            <ChatFinderPage
-                // eslint-disable-next-line react/jsx-props-no-spreading
-                {...args}
-                navigation={args.navigation}
-            />
+        <ComposeProviders components={[OnyxProvider, LocaleContextProvider, KeyboardStateProvider]}>
+            <OptionListContextProvider>
+                <ChatFinderPage
+                    // eslint-disable-next-line react/jsx-props-no-spreading
+                    {...args}
+                    navigation={args.navigation}
+                />
+            </OptionListContextProvider>
         </ComposeProviders>
     );
 }
@@ -158,7 +157,7 @@ function ChatFinderPageWithCachedOptions(args: ChatFinderPageProps) {
     );
 }
 
-test('[Search Page] should render list with cached options', async () => {
+test('[ChatFinderPage] should render list with cached options', async () => {
     const {addListener} = TestHelper.createAddListenerMock();
 
     const scenario = async () => {
@@ -182,7 +181,7 @@ test('[Search Page] should render list with cached options', async () => {
     );
 });
 
-test('[Search Page] should interact when text input changes', async () => {
+test('[ChatFinderPage] should interact when text input changes', async () => {
     const {addListener} = TestHelper.createAddListenerMock();
 
     const scenario = async () => {
@@ -211,93 +210,3 @@ test('[Search Page] should interact when text input changes', async () => {
     );
 });
 
-test.skip('[Search Page] should render selection list', async () => {
-    const {triggerTransitionEnd, addListener} = TestHelper.createAddListenerMock();
-    const smallMockedPersonalDetails = getMockedPersonalDetails(5);
-
-    const scenario = async () => {
-        await screen.findByTestId('ChatFinderPage');
-        await waitFor(triggerTransitionEnd as Awaited<() => Promise<void>>);
-        await screen.findByTestId('selection-list');
-        await screen.findByText(smallMockedPersonalDetails['1'].login as TextMatch);
-        await screen.findByText(smallMockedPersonalDetails['2'].login as TextMatch);
-    };
-
-    const navigation = {addListener};
-
-    return (
-        waitForBatchedUpdates()
-            .then(() =>
-                Onyx.multiSet({
-                    ...mockedReports,
-                    [ONYXKEYS.PERSONAL_DETAILS_LIST]: smallMockedPersonalDetails,
-                    [ONYXKEYS.BETAS]: mockedBetas,
-                    [ONYXKEYS.IS_SEARCHING_FOR_REPORTS]: true,
-                }),
-            )
-            // @ts-expect-error TODO: Remove this once TestHelper (https://github.com/Expensify/App/issues/25318) is migrated to TypeScript.
-            .then(() => measurePerformance(<ChatFinderPageWrapper navigation={navigation} />, {scenario}))
-    );
-});
-
-test('[Search Page] should search in selection list', async () => {
-    const {triggerTransitionEnd, addListener} = TestHelper.createAddListenerMock();
-
-    const scenario = async () => {
-        await screen.findByTestId('ChatFinderPage');
-        await waitFor(triggerTransitionEnd as Awaited<() => Promise<void>>);
-
-        const input = screen.getByTestId('selection-list-text-input');
-        const searchValue = mockedPersonalDetails['88'].login;
-
-        fireEvent.changeText(input, searchValue);
-        await screen.findByText(searchValue as TextMatch);
-    };
-
-    const navigation = {addListener};
-
-    return (
-        waitForBatchedUpdates()
-            .then(() =>
-                Onyx.multiSet({
-                    ...mockedReports,
-                    [ONYXKEYS.PERSONAL_DETAILS_LIST]: mockedPersonalDetails,
-                    [ONYXKEYS.BETAS]: mockedBetas,
-                    [ONYXKEYS.IS_SEARCHING_FOR_REPORTS]: true,
-                }),
-            )
-            // @ts-expect-error TODO: Remove this once TestHelper (https://github.com/Expensify/App/issues/25318) is migrated to TypeScript.
-            .then(() => measurePerformance(<ChatFinderPageWrapper navigation={navigation} />, {scenario}))
-    );
-});
-
-test('[Search Page] should click on list item', async () => {
-    const {triggerTransitionEnd, addListener} = TestHelper.createAddListenerMock();
-
-    const scenario = async () => {
-        await screen.findByTestId('ChatFinderPage');
-        const input = screen.getByTestId('selection-list-text-input');
-        await waitFor(triggerTransitionEnd as Awaited<() => Promise<void>>);
-
-        const searchValue = mockedPersonalDetails['4'].login as TextMatch;
-        fireEvent.changeText(input, searchValue);
-
-        const optionButton = await screen.findByText(searchValue);
-        fireEvent.press(optionButton);
-    };
-
-    const navigation = {addListener};
-    return (
-        waitForBatchedUpdates()
-            .then(() =>
-                Onyx.multiSet({
-                    ...mockedReports,
-                    [ONYXKEYS.PERSONAL_DETAILS_LIST]: mockedPersonalDetails,
-                    [ONYXKEYS.BETAS]: mockedBetas,
-                    [ONYXKEYS.IS_SEARCHING_FOR_REPORTS]: true,
-                }),
-            )
-            // @ts-expect-error TODO: Remove this once TestHelper (https://github.com/Expensify/App/issues/25318) is migrated to TypeScript.
-            .then(() => measurePerformance(<ChatFinderPageWrapper navigation={navigation} />, {scenario}))
-    );
-});
